feature: improvements (#5)
* build: remove redundant dependencies

* doc: improve README

* refactor: move function

* doc: added TODO comment to metadata

* doc: fix typo

* fix: remove redundant importat

* fix: tests

* fix: wrong import path

* doc: added related task id

* doc: build metadata placeholder


* chore: naming

* build: check yarn lock files

* build: reintroduce hardhat-upgrades

---------
