+++ packages/contracts/test/10_unit-testing/11_plugin.ts
@@ -1,3 +1,4 @@
+import {createDaoProxy} from '../20_integration-testing/test-helpers';
 import {PLUGIN_CONTRACT_NAME} from '../../plugin-settings';
 import {
   Admin,
@@ -22,7 +23,7 @@ import {
   getInterfaceId,
   DAO_PERMISSIONS,
 } from '@aragon/osx-commons-sdk';
-import {DAO, DAOEvents, DAOStructs, DAO__factory} from '@aragon/osx-ethers';
+import {DAO, DAOEvents, DAOStructs} from '@aragon/osx-ethers';
 import {loadFixture} from '@nomicfoundation/hardhat-network-helpers';
 import {SignerWithAddress} from '@nomiclabs/hardhat-ethers/signers';
 import {expect} from 'chai';
@@ -430,32 +431,4 @@ async function fixture(): Promise<FixtureResult> {
     dummyActions,
     dummyMetadata,
   };
-}
-
-// TODO Move into OSX commons ?
-export async function createDaoProxy(
-  deployer: SignerWithAddress,
-  dummyMetadata: string
-): Promise<DAO> {
-  const daoImplementation = await new DAO__factory(deployer).deploy();
-  const daoProxyFactory = await new ProxyFactory__factory(deployer).deploy(
-    daoImplementation.address
-  );
-
-  const daoInitData = daoImplementation.interface.encodeFunctionData(
-    'initialize',
-    [
-      dummyMetadata,
-      deployer.address,
-      ethers.constants.AddressZero,
-      dummyMetadata,
-    ]
-  );
-  const tx = await daoProxyFactory.deployUUPSProxy(daoInitData);
-  const event = await findEvent<ProxyCreatedEvent>(
-    tx,
-    daoProxyFactory.interface.getEvent('ProxyCreated').name
-  );
-  const dao = DAO__factory.connect(event.args.proxy, deployer);
-  return dao;
 }
